chore(docs): re-enable link checking
The dangling links being reported by the link checking processor were actually
valid warnings.  This commit re-enables the link checker and fixes those dangling
links.

- the project is not using componentGroups so this processor is disabled.
- the project is not using module pages but the ones being generated were using a
template that was appropriate for the angular.js project, so it had invalid
content as well as links
- the index.html template contained an unnecessary invalid href, since the link
is being handled by an ng-click handler.

**For future reference, it is easier to disable a processor by simply setting
its `$enabled` property to `false` - this works for all processors**