refactor(button): replace with an anchor or button
Closes #423. Closes #508. Closes #230. Closes #558.

BREAKING CHANGE: When using the `<md-button>` directive, the compiled
element will now be a normal `<a> or `<button>` element with the class
`md-button` on it. There will no longer be `<md-button>` elements in
your final DOM.

This means that any css referencing the `md-button` element selector
should now reference the `.md-button` class selector.

Change your code from this:

```css
md-button {
  color: red;
}
```

To this:

```css
.md-button {
  color: red;
}
```