fix(demo,swipe): add missing demo for swipe
Fixes an issue where the link on the toolbar redirects to '/' since no
demo exists.  One solution was to hide the link, but the process to
determine if an ngDoc has a demo would be more complicated.  Additional
demos can be added in the future to show more advanced usage.

Closes #2813x