fix(docs): fix demos GitCDN link to use proper URL for versioning
The GitCDN link used in the current demos points to a URL which
is redirected to a new one. When this redirect occurs, the
version info supplied in the URL is lost and the file always
points to HEAD making it difficult to track down.

Fix snapshot script to use the new link which retains the
version information.

Closes #4922.