docs: Fix attribute directives in dropdown.
Currently the attribute directives in the API Reference dropdown
available on all of the demo pages appear as element directives
because they are all wrapped with `<>` instead of `[]`. Add the
necessary code to change this and update dgeni to latest version
with necessary additions.

Also changes `<md-whiteframe>` to be an element or an attribute
directive as the examples show.

Fixes #7620.

Closes #8571