fix(autocomplete): remove unnecessary execution
Currently when the `selectedItem` changes, we check for the `searchText` being changed as well.
This requires querying the `getDisplayValue` function with the current `searchText`.

When the `searchText` is empty, we could remove that unnecessary execution and just skip the comparison (Reduces queries)

Also this improves the compatibility with the `virtual repeater`, since the repeater changes the scope bindings during runtime.
e.g `searchText` will be set to `undefined` and a ignorable error will be thrown.

Closes #9221