docs: Fix many links and other issues. (#9360)
There are a few places in the docs that point to `/something/somewhere`
which works on localhost, but fails on the live site because the live
site adds `/latest` or `/HEAD` to the URL.

 - Update: Absolute links to be relative.
 - Update: Material Design links to use new url (material.goole.com) to
   avoid redirects (and potentially broken links in the future).

Additionally, fix a few bugs with some of our doc-specific code.

 - Fix: Null pointer exception when sidenav is closed (i.e. on mobile).
 - Fix: `<hljs>` code snippets removed useful whitespace.

Additionally, update the Getting Started page with formatting/styling
improvements and new links.

 - Update: Formatting/styling/extra tags.
 - New: Add new links for Material Start.
 - Fix: Link titles all say `Link opens in a new window`.

Finally, fix various typos/formatting issues including (but not limited
to):

 - Update: Notes should look the same across pages.
 - Fix: Incorrect markdown support in panel docs.
 - Update: Add `<hljs>` snippet formatting to panel docs.
 - New: Add description to `$mdAriaProvider.disableWarnings()` method.

   This was mainly because it visually looked odd not to have one.

Fixes #9352. Related to #9357.