fix(datepicker): jumping forward if min date is in the same month as model (#9305)
Fixes the datepicker jumping forward if the current month is also the first one being rendered (e.g. when the min date is in the same month as the model). This was caused by an extra unnecessary call to `generateContent` which made the first month in the calendar to be re-rendered at the bottom after the user has scrolled past it. When the first month is at the bottom, the calendar tries to focus the active date, causing the browser to scroll to the bottom. Afterward the virtual repeater updates the view again, which makes the calendar skip a few months.

Fixes #9284.