feat(panel): Panel grouping (#9538)
Panels are now capable of being grouped together by any designation
the user sees fit. Grouping allows the user to specify logic for a
collection of panels. For instance, all popups in the header can be
considered one group with a maximum of one popup open at a time.
Dialogs within dialogs are another example of groups. This
implementation prevents the panel from storing a ridiculous amount
of data in memory or polluting the DOM with several panels. This
helps performance and gives the user more flexibility about how to
use panels.

Fixes #8971