update(build): sort panel styles to the top (#9854)
Since the CSS is being compiled in alphabetical order, the mdPanel styles end up after most components in the compiled CSS, which makes it hard for components that implement mdPanel to override it's styles (the `opacity` in particular). This change sorts the panel styles to the top, right after all of the core structural CSS. This is a temporary solution until we get rid of interimElement and move mdPanel into the core.