fix(virtualRepeat): DOM manipulation may alter scroll (#10177)
Fixes [#10144](https://github.com/angular/material/issues/10144) - There is an issue with Chrome that when the browser window is small enough, the intial DOM manipulations alter the scroll, which leads to the caldendar opening to the wrong month. Scrolling twice on the first render to ensure that the position is correct fixes this.